Remove TODO about ensemble accuracy
There isn't actually any stage where we ask the _ensemble_ for
predictions, and therefore no stage where the ensemble votes.

What _do_ we ask the ensemble for? Rewards. That's it.

Where _do_ predictions come in? We never actually query predictions from
even individual ensemble members. The prediction outputs of the network
are only used for training.